requirements.txt
Django==3.2.5
djangorestframework==3.13.1
drf-yasg==1.20.0